% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.chr.preventable.hospitalizations-package.R
\docType{package}
\name{dc.chr.preventable.hospitalizations-package}
\alias{dc.chr.preventable.hospitalizations}
\alias{dc.chr.preventable.hospitalizations-package}
\title{dc.chr.preventable.hospitalizations: Social Data Commons: Health, Behavior, Diet: Preventable Hospitalizations}
\description{
Allows user to easily get Virginia preventable hospitalization data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
